+++ src/pages/EditProfile.jsx
@@ -1,8 +1,8 @@
-import { onAuthStateChanged, updateProfile } from "firebase/auth";
-import { useEffect, useState } from "react";
+import { onAuthStateChanged } from "firebase/auth";
+import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import Button from "src/components/Button";
 import ImageChange from "src/components/EditProFileComponenets/ImageChange";
+import NameChange from "src/components/EditProFileComponenets/NameChange";
 import HomeHeader from "src/components/HomeComponents/HomeHeader";
 import Loading from "src/components/Loading";
 import { auth } from "src/firebase";
@@ -12,7 +12,6 @@ import styled from "styled-components";
 
 function EditProfile() {
   const { user } = useSelector((state) => state.users);
-  const [name, setName] = useState("");
   const dispatch = useDispatch();
   const dispatchUser = () => {
     dispatch(log(user));
@@ -23,18 +22,6 @@ function EditProfile() {
       dispatch(log(user));
     });
   }, [dispatch]);
-
-  const proFileChange = async () => {
-    if (name === "") {
-      alert("이름을 입력해주시기 바랍니다.");
-      return;
-    }
-    await updateProfile(auth.currentUser, {
-      displayName: name,
-    });
-    dispatchUser();
-    setName("");
-  };
 
   if (user === null)
     return (
@@ -50,20 +37,7 @@ function EditProfile() {
       <HomeHeader />
       <Background>
         <ImageChange user={user} dispatchUser={dispatchUser} />
-        <NameChange>
-          <p>Name Change</p>
-          <NameChangeInput
-            type="text"
-            placeholder="변경할 이름을 입력해주세요"
-            value={name}
-            onChange={(event) => setName(event.target.value)}
-          />
-          <Button
-            content={"Name Change"}
-            width={"110"}
-            onClick={proFileChange}
-          />
-        </NameChange>
+        <NameChange dispatchUser={dispatchUser} />
       </Background>
     </>
   );
@@ -86,21 +60,4 @@ const Background = styled.div`
   justify-content: center;
   align-items: center;
   margin-top: 50px;
-`;
-
-const NameChange = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 30px;
-  margin-top: 50px;
-`;
-
-const NameChangeInput = styled.input`
-  height: 40px;
-  width: 300px;
-  border: none;
-  box-shadow: 1px 3px 5px 5px gray;
-  border-radius: 12px;
-  padding: 5px;
 `;
